requirements.txt
Flask
pymysql
mysql-connector-python
pandas==1.3.5
psython
bcrypt